requirements.txt
pandas
streamlit
streamlit_option_menu
psycopg2-binary
pandas
boto3